Fix for a problem on updating receipt information
Fix a problem when updating receipt information. When a receipt was
saved it lost it pictures. Server sended a picture object with a empty
receipt array as a response and that ended up to messing UI state.

The user interface ended up losing all receipt pictures after saving
receipt information.